+++ src/i18next.js
@@ -42,15 +42,13 @@ class I18n extends EventEmitter {
       options = {};
     }
 
-    // temporal backwards compatibility WHITELIST REMOVAL
-    if (options.whitelist && !options.supportedLngs) {
-      this.logger.deprecate('whitelist', 'option "whitelist" will be renamed to "supportedLngs" in the next major - please make sure to rename this option asap.');
+    if (!options.defaultNS && options.ns) {
+      if (typeof options.ns === 'string') {
+        options.defaultNS = options.ns;
+      } else {
+        options.defaultNS = options.ns[0];
+      }
     }
-    if (options.nonExplicitWhitelist && !options.nonExplicitSupportedLngs) {
-      this.logger.deprecate('whitelist', 'options "nonExplicitWhitelist" will be renamed to "nonExplicitSupportedLngs" in the next major - please make sure to rename this option asap.');
-    }
-    // end temporal backwards compatibility WHITELIST REMOVAL
-
 
     this.options = { ...getDefaults(), ...this.options, ...transformOptions(options) };
 
@@ -262,10 +260,25 @@ class I18n extends EventEmitter {
     const deferred = defer();
     this.emit('languageChanging', lng);
 
+    const setLngProps = (l) => {
+      this.language = l;
+      this.languages = this.services.languageUtils.toResolveHierarchy(l);
+      // find the first language resolved languaged
+      this.resolvedLanguage = undefined;
+      if (['cimode', 'dev'].indexOf(l) > -1) return;
+      for (let li = 0; li < this.languages.length; li++) {
+        const lngInLngs = this.languages[li];
+        if (['cimode', 'dev'].indexOf(lngInLngs) > -1) continue;
+        if (this.store.hasLanguageSomeTranslations(lngInLngs)) {
+          this.resolvedLanguage = lngInLngs;
+          break;
+        }
+      }
+    };
+
     const done = (err, l) => {
       if (l) {
-        this.language = l;
-        this.languages = this.services.languageUtils.toResolveHierarchy(l);
+        setLngProps(l);
         this.translator.changeLanguage(l);
         this.isLanguageChangingTo = undefined;
         this.emit('languageChanged', l);
@@ -286,8 +299,7 @@ class I18n extends EventEmitter {
 
       if (l) {
         if (!this.language) {
-          this.language = l;
-          this.languages = this.services.languageUtils.toResolveHierarchy(l);
+          setLngProps(l);
         }
         if (!this.translator.language) this.translator.changeLanguage(l);
 
@@ -359,7 +371,7 @@ class I18n extends EventEmitter {
       return false;
     }
 
-    const lng = this.languages[0];
+    const lng = this.resolvedLanguage || this.languages[0];
     const fallbackLng = this.options ? this.options.fallbackLng : false;
     const lastLng = this.languages[this.languages.length - 1];
 
@@ -433,7 +445,7 @@ class I18n extends EventEmitter {
   }
 
   dir(lng) {
-    if (!lng) lng = this.languages && this.languages.length > 0 ? this.languages[0] : this.language;
+    if (!lng) lng = this.resolvedLanguage || (this.languages && this.languages.length > 0 ? this.languages[0] : this.language);
     if (!lng) return 'rtl';
 
     const rtlLngs = [
@@ -539,7 +551,8 @@ class I18n extends EventEmitter {
       options: this.options,
       store: this.store,
       language: this.language,
-      languages: this.languages
+      languages: this.languages,
+      resolvedLanguage: this.resolvedLanguage
     };
   }
 }