+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 21.0.0 (unreleased)
+
+**This is a major breaking release:**
+
+- introduce new v4 json format (intl plural forms), simplifyPluralSuffix not used anymore
+- `skipOnVariables` by default now is true
+- automatically detect natural language keys (no need to set nsSeparator or keySeparator to false)
+- remove deprecated whitelist features
+- introduce new i18next.resolvedLanguage property
+
 ### 20.6.1
 
 - fix deepFind if looking for nesting array item